refactor: share introspection api types
