+++ frontend/src/api/introspection.ts
@@ -1,17 +1,7 @@
 import { apiFetch } from '@/api/client';
+import type { IntrospectionResponse } from '@/types/introspection';
 
-export type IntrospectionReport = {
-  generatedAt: string;
-  summary: string;
-  signals: Array<{ id: string; status: string; detail: string }>;
-  recommendations: string[];
-};
-
-export type IntrospectionResponse = {
-  ok: boolean;
-  awarenessScore: number;
-  introspectionReport: IntrospectionReport;
-};
+export type { IntrospectionReport, IntrospectionResponse } from '@/types/introspection';
 
 export async function fetchIntrospectionReport(): Promise<IntrospectionResponse> {
   return apiFetch<IntrospectionResponse>('/genesis/introspection-report');
