+++ frontend/src/types/introspection.ts
@@ -1,0 +1,12 @@
+export type IntrospectionReport = {
+  generatedAt: string;
+  summary: string;
+  signals: Array<{ id: string; status: string; detail: string }>;
+  recommendations: string[];
+};
+
+export type IntrospectionResponse = {
+  ok: boolean;
+  awarenessScore: number;
+  introspectionReport: IntrospectionReport;
+};
